Move things API into lib
